demo_backend/requirements_demo.txt
# Demo-only dependencies (install with pip)
pillow>=10.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
streamlit>=1.33.0
